refactor: extracts common methods for MemorandoUtil
